chore: resolve circular depends
