+++ src/ui/index.ts
@@ -1,3 +1,3 @@
 export * from "./pages";
 export * from "./layouts";
-export { Loading, ModalPortal, StandaloneErrorBoundary } from "./shared";
+export { Loading, ModalPortal } from "./shared";
