+++ src/ui/layouts/index.ts
@@ -8,3 +8,4 @@ export * from "./database-detail-layout";
 export * from "./environment-detail-layout";
 export * from "./hero-bg-layout";
 export * from "./unauth-required";
+export * from "./error-boundary";
